mcp-fantasy-draft/agent-yahooff/pyproject.toml
[project]
name = "yahoo-fantasy-football-agent"
version = "0.1.0"
description = "MCP agent for Yahoo Fantasy Football - interact with your team and waiver wire via Copilot CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
    "yahoo-fantasy-api>=2.12.0",
    "yahoo-oauth>=2.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
